tests/wpt/tests/css/css-conditional/container-queries/scroll-state/at-container-overflowing-parsing.html

<!DOCTYPE html>
<title>CSS Conditional Test: @container overflowing query parsing</title>
<link rel="help" href="https://drafts.csswg.org/css-conditional-5/#container-rule">
<link rel="help" href="https://drafts.csswg.org/css-conditional-5/#overflowing">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/css-conditional/container-queries/support/cq-testcommon.js"></script>
<div style="container-name:name;container-type:scroll-state">
  <main id="cq-main"></main>
</div>
<script>
  setup(() => assert_implements_scroll_state_container_queries());

  test_cq_condition_known('scroll-state(overflowing)');
  test_cq_condition_known('scroll-state(overflowing: none)');
  test_cq_condition_known('scroll-state(overflowing: top)');
  test_cq_condition_known('scroll-state(overflowing: left)');
  test_cq_condition_known('scroll-state(overflowing: bottom)');
  test_cq_condition_known('scroll-state(overflowing: right)');
  test_cq_condition_known('scroll-state(overflowing: inline-start)');
  test_cq_condition_known('scroll-state(overflowing: inline-end)');
  test_cq_condition_known('scroll-state(overflowing: block-start)');
  test_cq_condition_known('scroll-state(overflowing: block-end)');
  test_cq_condition_known('(scroll-state(overflowing: inline-end))');
  test_cq_condition_known('scroll-state((overflowing: left))');
  test_cq_condition_known('scroll-state(not ((overflowing: bottom) and (overflowing: right)))');
  test_cq_condition_known('scroll-state((overflowing: left) or (overflowing: top))');

  test_cq_condition_unknown('scroll-state(overflowing: auto)');
  test_cq_condition_unknown('scroll-state(overflowing: true)');
  test_cq_condition_unknown('scroll-state(style(overflowing: left))');
  test_cq_condition_unknown('style(scroll-state(overflowing: left))');
  test_cq_condition_unknown('scroll-state(overflowing:)');
</script>
